+++ spec/rubocop/cop/style/symbol_array_spec.rb
@@ -14,6 +14,21 @@ describe RuboCop::Cop::Style::SymbolArray, :config do
       expect(cop.messages).to eq(['Use `%i` or `%I` for an array of symbols.'])
     end
 
+    it 'autocorrects arrays of symbols', ruby: 2 do
+      new_source = autocorrect_source(cop, '[:one, :two, :three]')
+      expect(new_source).to eq('%i(one two three)')
+    end
+
+    it 'uses %I when appropriate', ruby: 2 do
+      new_source = autocorrect_source(cop, '[:"\\t", :"\\n", :three]')
+      expect(new_source).to eq('%I(\\t \\n three)')
+    end
+
+    it "doesn't break when a symbol contains )", ruby: 2 do
+      new_source = autocorrect_source(cop, '[:one, :")", :three]')
+      expect(new_source).to eq('%i(one \\) three)')
+    end
+
     it 'does not reg an offense for array with non-syms', ruby: 2 do
       inspect_source(cop, '[:one, :two, "three"]')
       expect(cop.offenses).to be_empty
@@ -52,6 +67,11 @@ describe RuboCop::Cop::Style::SymbolArray, :config do
       inspect_source(cop, '%i(one two three)')
       expect(cop.offenses.size).to eq(1)
       expect(cop.messages).to eq(['Use `[]` for an array of symbols.'])
+    end
+
+    it 'autocorrects an array starting with %i', ruby: 2 do
+      new_source = autocorrect_source(cop, '%i(one two three)')
+      expect(new_source).to eq('[:one, :two, :three]')
     end
   end
 end